# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-redable
        type: string
    type: object
  internal_features_statistics_transport_http.GetStatisticsResponse:
    properties:
      tasks_average_completed_time:
        example: 2m10s
        type: string
      tasks_completed:
        example: 10
        type: integer
      tasks_completed_rate:
        example: 50
        type: number
      tasks_created:
        example: 20
        type: integer
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        example: 2
        type: integer
      discription:
        example: Сделать....
        maxLength: 1000
        minLength: 1
        type: string
      title:
        example: Сходить на тренировку
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: 2
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T12:00:00Z"
        type: string
      description:
        example: Сделать...
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Сходить на тренировку
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: 2
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T12:00:00Z"
        type: string
      description:
        example: Сделать...
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Сходить на тренировку
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        type: boolean
      description:
        example: "null"
        type: string
      title:
        example: Сходить на тренировку
        type: string
    type: object
  internal_features_tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        example: 2
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T12:00:00Z"
        type: string
      description:
        example: Сделать...
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Сходить на тренировку
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: 2
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T12:00:00Z"
        type: string
      description:
        example: Сделать...
        type: string
      id:
        example: 1
        type: integer
      title:
        example: Сходить на тренировку
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Jetty Junk
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+79998887766"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Jetty Junk
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Jetty Junk
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Jetty Junk
        type: string
      phone_number:
        example: "+79998887766"
        type: string
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Jetty Junk
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Jetty Junk
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: localhost:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
  /statistics:
    get:
      consumes:
      - application/json
      description: Получение статистики по задачам с опциональной фильтрацией по user_id
        и/или временному промежутку
      parameters:
      - description: Фильтрация статистики по конкретному пользователю
        in: query
        name: id
        type: integer
      - description: 'Начало промежутка рассмотрения статистике (включительно), формат:
          YYYY-MM-DD'
        in: query
        name: from
        type: string
      - description: 'Конец промежутка рассмотрения статистике (не включительно),
          формат: YYYY-MM-DD'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешно получение статистики
          schema:
            $ref: '#/definitions/internal_features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение статистики
      tags:
      - statistics
  /tasks:
    get:
      description: Просмотр списка существующих задач с опциональной пагинацией и/или
        фильтрации по ID автора задачи
      parameters:
      - description: Фильтрация задач по ID автора
        in: query
        name: user_id
        type: integer
      - description: Размер старницы с задачами
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с задачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка задач
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу у пользователя
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Задача успешно созданна
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удалить соществующею задачу по её ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление задачи
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Получение существующей задачи в системе по её ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение задачи
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение задачи
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Измение информации об уже существующей задаче
        ### Логика обновления полей (Theree-state logic):
        1.**Поле не передано**: `description` игнорируется, значение в БД не меняется
        2.**Переданно значение**: `description` "`Сделать 10 отжиманий`" - устанавливает новый `description` в БД
        3.**Переданн null**: `"description":null` - очищает поле в БД
        Ограничение `title` не может быть как null
      parameters:
      - description: ID изменяемой задачи
        in: path
        name: id
        required: true
        type: integer
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённая задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
  /users:
    get:
      description: Просмотр списка существующих пользователей с опциональной пагинацией
      parameters:
      - description: Размер старницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаление существующего пользователя в системе по его ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление пользователя
      tags:
      - users
    get:
      description: Получение существующего пользователя в системе по его ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение пользователя
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Измение информации об уже существующем в системе пользователя
        ### Логика обновления полей (Theree-state logic):
        1.**Поле не передано**: `phone_number` игнорируется, значение в БД не меняется
        2.**Переданно значение**: `phone_number` "`+71112223344`" - устанавливает новый `phone_number` в БД
        3.**Переданн null**: `"phone_number":null` - очищает поле в БД
        Ограничение `full_name` не может быть как null
      parameters:
      - description: ID изменяемого пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённый пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_jettyjunk_goland-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Изменения пользователя
      tags:
      - users
swagger: "2.0"