# internal/features/tasks/transport/http/create_task.go
package tasks_transport_http

import (
	"net/http"

	"github.com/jettyjunk/goland-todoapp/internal/core/domain"
	core_logger "github.com/jettyjunk/goland-todoapp/internal/core/logger"
	core_http_request "github.com/jettyjunk/goland-todoapp/internal/core/transport/http/request"
	core_http_response "github.com/jettyjunk/goland-todoapp/internal/core/transport/http/response"
)

type CreateTaskRequest struct {
	Title        string  `json:"title" validate:"required,min=1,max=100" example:"Сходить на тренировку"`
	Discription  *string `json:"discription" validate:"omitempty,min=1,max=1000" example:"Сделать...."`
	AuthorUserID int     `json:"author_user_id" validate:"required" example:"2"`
}

type CreateTaskResponse TaskDTOResponse

// CreateTask 	godoc
// @Summary 	Создать задачу
// @Description Создать новую задачу у пользователя
// @Tags 		tasks
// @Accept 		json
// @Produce 	json
// @Param 		request body 		CreateTaskRequest true "CreateTask тело запроса"
// @Success 	201 	{object} 	CreateTaskResponse "Задача успешно созданна"
// @Failure		400 	{object} 	core_http_response.ErrorResponse "Bad Request"
// @Failure 	404 	{object}	core_http_response.ErrorResponse "Author not found"
// @Failure 	500 	{object} 	core_http_response.ErrorResponse "Internal server error"
// @Router 		/tasks 	[post]
func (h *TasksHTTPHandler) CreatedTask(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, w)

	var request CreateTaskRequest
	if err := core_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(err, "failde to decode and validate HTTP request")
		return
	}

	taskDomain := domain.NewTaskUnitialized(request.Title, request.Discription, request.AuthorUserID)
	taskDomain, err := h.tasksService.CreateTask(ctx, taskDomain)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to created task")
		return
	}

	response := taskDTOFromDomain(taskDomain)

	responseHandler.JsonResponse(response, http.StatusCreated)
}